+++ Client/src/redux/store.ts
@@ -10,7 +10,7 @@ import { flowSelectionReducer } from "./flowSelection/flowSelectionSlice";
 import { classSearchReducer } from "./classSearch/classSearchSlice";
 import { sectionSelectionReducer } from "./sectionSelection/sectionSelectionSlice";
 import { scheduleReducer } from "./schedule/scheduleSlice";
-import { scheduleBuilderLogReducer } from "./scheduleBuilderLog/scheduleBuilderLog";
+import { scheduleBuilderLogReducer } from "./scheduleBuilderLog/scheduleBuilderLogSlice";
 import { scheduleBuilderChatReducer } from "./scheduleBuilderChat/scheduleBuilderChatSlice";
 
 const store = configureStore({
